add call_rcu support
Unlike the RCU code, this is not (yet) meant to be replaced by liburcu,
because we want to run call_rcu callbacks in the main thread, so that
they run under the global lock.

The data structure is indeed based on those found in liburcu (this makes
me feel safer), but adapted to replace futexes with QemuEvents.

This has a dependency on the iothread, because it relies on the fact that
qemu_schedule_bh works outside the global lock.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>