lockless access to aiocb->ret and aiocb->queued
aiocb->state can be accessed lockless if we leave the removal from
the request_list to the thread-pool to the worker threads.

With two concurrent cancellations, either both will block waiting
for completion, or one of them will fail the cmpxchg and see
THREAD_CANCELLED or THREAD_DONE.  It will then exit immediately.

Drawback: if the pool is full, cancelling new requests will have
to wait for one request to complete.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>