block: clear dirty bitmap when discarding
Note that resetting bits in the dirty bitmap is done _before_ actually
processing the request.  This way, when there are concurrent write and
discard requests, the outcome will always be that the blocks are
marked dirty.  This scenario should never happen, but it is safer
to do it this way.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>