qemu-iotests: add mirroring test case
The tests are not meant to offer full coverage; in particular, there is
no concurrent I/O going on from the guest.  However, especially in the
case of blkdebug-based tests (introduced later in the series) they do
cover some paths that will usually be skipped by integration tests,
for example tests running on kvm-autotest.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>