9pfs-synth: use regular mutex
Even with the fix in the previous patch, the lockless handling of paths
in 9pfs-synth is wrong.  Paths can outlive rcu_read_unlock arbitrarily
via the V9fsPath objects that 9pfs-synth creates.  This would require
a reference counting mechanism that is not there and is quite hard to
retrofit.

It seems to me that this was a premature optimization, so replace
everything with a simple mutex.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>