build: avoid duplicate appearance of files
cutils.c and iov.c appeared in the Makefile in multiple spots.  Clean
this up by moving them respectively to oslib-obj-y and block-obj-y.
Notably, qemu-ga and user-mode emulation both need cutils.c but not iov.c.

async.o was linked in tools via both tools-obj-y and block-obj-y.
Remove the former, which is unnecessary.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>