migration: fix rate limiting
buffered_rate_limit is called to prevent the RAM migration callback
from putting too much data in the buffer.  So it has to check against
the amount of data currently in the buffer, not against the amount
of data that has been transferred so far.

s->bytes_xfer is used to communicate between successive calls of
buffered_put_buffer, with buffered_rate_tick resetting it every
now and then to prevent moving too much buffered data to the socket
at once.  But it does not matter for the producer of the data.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>